create model folder to store model objects
