+++ FinalExamPurchaseOrderManagement/FinalExamPurchaseOrderManagement/BussinessLogic/Model/POHead.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace FinalExamPurchaseOrderManagement.BussinessLogic.Model
+{
+    public class POHead
+    {
+        public string SupplierCode { get; set; }
+        public string SupplierName { get; set; }
+        public string StockSiteCode { get; set; }
+        public string StockSiteName { get; set; }
+        public DateTime? OrderDate { get; set; }
+        public string Country { get; set; }
+        public string Note { get; set; }
+        public string Address { get; set; }
+        public string PostCode { get; set; }
+    }
+}
