+++ FinalExamPurchaseOrderManagement/FinalExamPurchaseOrderManagement/FinalExamPurchaseOrderManagement.csproj
@@ -128,6 +128,9 @@
     <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
+    <Compile Include="BussinessLogic\Model\POHead.cs" />
+    <Compile Include="BussinessLogic\Model\POLine.cs" />
+    <Compile Include="BussinessLogic\Model\POList.cs" />
     <Compile Include="BussinessLogic\POService\POService.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Global.asax.cs">
